analice: parse go.mod instead of go.sum
go.sum contained test dependencies of dependencies, they are not vendored and therefore should not be part of NOTICE (and related formats). analice does continue to work in those cases, but issues an ugly warning at runtime. The implementation is somewhat hacky, hoping that a parser for go.mod will be exported someday.